refactor: table group 관련 update method 생성 및 디렉토리 변경
