+++ src/main/java/kitchenpos/domain/ordertable/OrderTableRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.domain.ordertable;
 
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
